+++ src/components/Combobox/Common/index.ts
@@ -1,0 +1,2 @@
+export * from "./useCombobox";
+export * from "./ComboboxWrapper";
